Twitter/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
generics==3.4.0
jsonfield==3.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.1
sqlparse==0.4.2
swapper==1.1.2.post1
